refactor: introduce a base class for ElasticsearchDocumentStore (#5228)
* introduce a base class

* forgot

* fix linting

* try

* try

* schema generation doesnt support aliasing, use the same name
